+++ CHANGELOG.md
@@ -6,4 +6,4 @@ Release Please maintains this file from Conventional Commit messages.
 
 ## 0.1.0 - 2026-07-03
 
-- Initial Domoticz Import custom integration scaffold.
+- Initial Domoticz Sync custom integration scaffold.